refactor : 내부 클래스를 외부 패키지로 이동한다.
